+++ src/Model1.hpp
@@ -32,7 +32,7 @@ public:
   void AlignRandomly();
 
   // run one iteration of Gibbs sampling over the corpus
-  void RunIteration(double temp);
+  void RunIteration(float temp);
 
   // get counts and joint counts, used as initial parameters by
   // subsequent models
